fixes to mobile illo
fixing the mobile illo!

the h2 in the .intro-type had a width set, not a max-width. this was causing the issues where the type was going too long, because it wasnt going smaller than 400px even when the screen was smaller!

illustration needed a background-size property set, which i did as 'cover' so it would show up. also took off the max-width on the illo so it stretches to the full width on wider screens!